+++ CHANGELOG.md
@@ -10,6 +10,7 @@ To see compatibility information, `xsvutils help compatibility`.
 - add --src, --dst options of inshour, insdate subcommand on v2
 - rename subcommand name `insdeltasec` to `inssecinterval`
 - remove `addlinenum2` subcommand
+- add `inscopy` subcommand on v2
 
 ## version 0.2.2 (2018-03-07)
 - add `inshour` subcommand on v2